Minor changes in Extractor to handle non-existing directories and add some more models that I tried

+++ featureextraction/src/main/java/tu/kn/ghrepoclassifier/featureextraction/Extractor.java
@@ -3,15 +3,14 @@ package tu.kn.ghrepoclassifier.featureextraction;
 import au.com.bytecode.opencsv.CSVWriter;
 import org.apache.commons.io.FileUtils;
 import tu.kn.ghrepoclassifier.Config;
-import tu.kn.ghrepoclassifier.featureextraction.features.text.ExtractText;
+import tu.kn.ghrepoclassifier.featureextraction.features.FeatureExtractionUnbiased;
 import tu.kn.ghrepoclassifier.serialization.Serializer;
 import tu.kn.ghrepoclassifier.serialization.data.RepoData;
 
 import java.io.*;
+import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.List;
-
-import static tu.kn.ghrepoclassifier.featureextraction.features.text.ExtractText.extractFeatures;
 
 /**
  * Created by felix on 04.12.16.
@@ -62,7 +61,8 @@ public class Extractor {
 					
 					if (cleaner.isValid(repo)) {
 						System.out.println(repo.getFull_name());
-						entries[1] = extractFeatures(repo);
+						// entries[1] = ExtractText.extractFeatures(repo);
+						entries[1] = FeatureExtractionUnbiased.extractFeatures(repo);
 						entries[0] = repo.getHtmlUrl().toString();
 
 						//System.out.println(entries[0] + ", " + entries[1]);
@@ -98,11 +98,15 @@ public class Extractor {
 	public static void extractFromDir(String inputDir, String outputDirS, boolean isTest) throws IOException {
 		File outputDir = new File(outputDirS);
 
-		FileUtils.cleanDirectory(outputDir);
+		if (Files.exists(outputDir.toPath()))
+			FileUtils.cleanDirectory(outputDir);
+		else
+			Files.createDirectories(outputDir.toPath());
 
 		extract(inputDir, outputDir.getAbsolutePath(), isTest);
 
-		List<String> labels = ExtractText.getFeatureLabels();
+		// List<String> labels = ExtractText.getFeatureLabels();
+		List<String> labels = FeatureExtractionUnbiased.getFeatureLabels();
 		createSchema(outputDir, labels);
 	}
 
@@ -111,12 +115,16 @@ public class Extractor {
 		String inputDir = Config.get("sample.generation.output.path");
 		String outputDir = Config.get("feature.extraction.output.path");
 
-		extractFromDir(inputDir, outputDir, false);
-
 		String inputDirAttachmentA = Config.get("attachmentA.download.output.path");
 		String outputDirAttachmentA = Config.get("attachmentA.feature.extraction.output.path");
 
+		String inputDirAttachmentB = Config.get("attachmentB.download.output.path");
+		String outputDirAttachmentB = Config.get("attachmentB.feature.extraction.output.path");
+
+		extractFromDir(inputDir, outputDir, false);
+
 		extractFromDir(inputDirAttachmentA, outputDirAttachmentA, true);
-		
+
+		extractFromDir(inputDirAttachmentB, outputDirAttachmentB, true);
 	}
 }
